% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_fish_doc.R
\docType{data}
\name{sbc_fish}
\alias{sbc_fish}
\title{Santa Barbara Channel fish abundance}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 74019 rows and 25 columns.
}
\source{
{Santa Barbara Coastal LTER and D. Reed. 2020. SBC LTER: Reef: Long-term experiment: Kelp removal: Fish abundance ver 19. Environmental Data Initiative. https://doi.org/10.6073/pasta/ecf2e269db7a4807bcaa765422d8186c (Accessed 2021-02-20).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-sbc.30.19}
}
\usage{
sbc_fish
}
\description{
abundance and size of reef-associated fish
}
\keyword{datasets}
